requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.38
ultralytics-thop==2.0.12
urllib3==2.2.3
tensorboard
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
black==24.10.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
detectron2==0.6
fonttools==4.55.0
fvcore==0.1.5.post20221221
grpcio==1.68.0
hydra-core==1.3.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mypy-extensions==1.0.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.29.0
pycocotools==2.0.8
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tomli==2.2.1
torch==1.10.0
torchvision==0.11.2
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
yacs==0.1.8
zipp==3.21.0
